requirements.txt
streamlit
numpy
pandas
scipy
matplotlib
Pillow